+++ libraries/INA228/INA228.cpp
@@ -1,6 +1,6 @@
 //    FILE: INA228.cpp
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.1.1
+// VERSION: 0.1.2
 //    DATE: 2024-05-09
 // PURPOSE: Arduino library for INA228 voltage, current and power sensor.
 //     URL: https://github.com/RobTillaart/INA228
@@ -31,6 +31,23 @@
 #define INA228_POWER_LIMIT          0x11    //  16   RW
 #define INA228_MANUFACTURER         0x3E    //  16   R-
 #define INA228_DEVICE_ID            0x3F    //  16   R-
+
+
+//  CONFIG MASKS (register 0)
+#define INA228_CFG_RST              0x8000
+#define INA228_CFG_RSTACC           0x4000
+#define INA228_CFG_CONVDLY          0x3FC0
+#define INA228_CFG_TEMPCOMP         0x0020
+#define INA228_CFG_ADCRANGE         0x0010
+#define INA228_CFG_RESERVED         0x000F
+
+
+//  ADC MASKS (register 1)
+#define INA228_ADC_MODE             0xF000
+#define INA228_ADC_VBUSCT           0x0E00
+#define INA228_ADC_VSHCT            0x01C0
+#define INA228_ADC_VTCT             0x0038
+#define INA228_ADC_AVG              0x0007
 
 
 ////////////////////////////////////////////////////////
@@ -76,60 +93,75 @@ uint8_t INA228::getAddress()
 float INA228::getBusVoltage()
 {
   uint32_t value = _readRegister(INA228_BUS_VOLTAGE, 3);
+  //  remove reserved bits.
   value >>= 4;
-  return value * 195.3125e-6;
+  float LSB = 195.3125e-6;  //  195.3125 uV
+  return value * LSB;
 }
 
 //  PAGE 25
 float INA228::getShuntVoltage()
 {
-  uint32_t value = _readRegister(INA228_SHUNT_VOLTAGE, 3);
-  value >>= 4;
-  //  depends on ADCRANGE in INA228_CONFIG register.
-  uint16_t config = _readRegister(INA228_CONFIG, 2);
-  if (config & 0x0008) return value * 78.125e-9;
-  return value * 312.5e-9;
+  //  LSB depends on ADCRANGE in INA228_CONFIG register.
+  float LSB = 312.5e-9;  //  312.5 nV
+  if (getADCRange() == 1)
+  {
+    LSB = 78.125e-9;     //  78.125 nV
+  }
+
+  //  remove reserved bits.
+  int32_t value = _readRegister(INA228_SHUNT_VOLTAGE, 3) >> 4;
+  //  handle negative values
+  if (value & 0x00800000)
+  {
+    value |= 0xFF000000;
+  }
+  return value * LSB;
 }
 
 //  PAGE 25 + 8.1.2
 float INA228::getCurrent()
 {
-  uint32_t value = _readRegister(INA228_CURRENT, 3);
-
-  //  PAGE 31 (8.1.2)
-  float shunt_cal = 13107.2e6 * _current_LSB * _shunt;
-  //  depends on ADCRANGE in INA228_CONFIG register.
-  uint16_t config = _readRegister(INA228_CONFIG, 2);
-  if (config & 0x0008) shunt_cal *= 4;
-  return value * shunt_cal;
+  //  remove reserved bits.
+  uint32_t value = _readRegister(INA228_CURRENT, 3) >> 4;
+  return value * _current_LSB;
 }
 
 //  PAGE 26 + 8.1.2
 float INA228::getPower()
 {
-  uint32_t value = _readRegister(INA228_POWER, 3);
+  uint32_t value = _readRegister(INA228_POWER, 3) ;
   //  PAGE 31 (8.1.2)
   return value * 3.2 * _current_LSB;
 }
 
-//  PAGE25
+//  PAGE 25
 float INA228::getTemperature()
 {
   uint32_t value = _readRegister(INA228_TEMPERATURE, 2);
-  return value * 7.8125e-6;
+  float LSB = 7.8125e-3;  //   milli degree Celsius
+  return value * LSB;
 }
 
 //  PAGE 26 + 8.1.2
-float INA228::getEnergy()
+double INA228::getEnergy()
 {
-  float value = _readRegisterF(INA228_ENERGY, 5);
+  //  read 40 bit unsigned as a double to prevent 64 bit ints
+  //  double might be 8 or 4 byte, depends on platform
+  //  40 bit ==> O(10^12)
+  double value = _readRegisterF(INA228_ENERGY, 5);
   //  PAGE 31 (8.1.2)
-  return value * 16 * 3.2 * _current_LSB;
+  return value * (16 * 3.2) * _current_LSB;
 }
 
-float INA228::getCharge()
+
+//  PAGE 26 + 8.1.2
+double INA228::getCharge()
 {
-  float value = _readRegisterF(INA228_CHARGE, 5);
+  //  read 40 bit unsigned as a float to prevent 64 bit ints
+  //  double might be 8 or 4 byte, depends on platform
+  //  40 bit ==> O(10^12)
+  double value = _readRegisterF(INA228_CHARGE, 5);
   //  PAGE 32 (8.1.2)
   return value * _current_LSB;
 }
@@ -142,7 +174,7 @@ float INA228::getCharge()
 void INA228::reset()
 {
   uint16_t value = _readRegister(INA228_CONFIG, 2);
-  value |= 0x8000;
+  value |= INA228_CFG_RST;
   _writeRegister(INA228_CONFIG, value);
 }
 
@@ -150,8 +182,8 @@ bool INA228::setAccumulation(uint8_t value)
 {
   if (value > 1) return false;
   uint16_t reg = _readRegister(INA228_CONFIG, 2);
-  reg &= ~0x4000;
-  if (value == 1) reg |= 0x4000;
+  reg &= ~INA228_CFG_RSTACC;
+  if (value == 1) reg |= INA228_CFG_RSTACC;
   _writeRegister(INA228_CONFIG, reg);
   return true;
 }
@@ -159,13 +191,13 @@ bool INA228::setAccumulation(uint8_t value)
 bool INA228::getAccumulation()
 {
   uint16_t value = _readRegister(INA228_CONFIG, 2);
-  return (value & 0x4000) > 0;
+  return (value & INA228_CFG_RSTACC) > 0;
 }
 
 void INA228::setConversionDelay(uint8_t steps)
 {
   uint16_t value = _readRegister(INA228_CONFIG, 2);
-  value &= ~0x3FC0;
+  value &= ~INA228_CFG_CONVDLY;
   value |= (steps << 6);
   _writeRegister(INA228_CONFIG, value);
 }
@@ -179,29 +211,29 @@ uint8_t INA228::getConversionDelay()
 void INA228::setTemperatureCompensation(bool on)
 {
   uint16_t value = _readRegister(INA228_CONFIG, 2);
-  value &= ~0x0020;
-  if (on) value |= 0x0020;
+  value &= ~INA228_CFG_TEMPCOMP;
+  if (on) value |= INA228_CFG_TEMPCOMP;
   _writeRegister(INA228_CONFIG, value);
 }
 
 bool INA228::getTemperatureCompensation()
 {
   uint16_t value = _readRegister(INA228_CONFIG, 2);
-  return value & 0x0020;
+  return (value & INA228_CFG_TEMPCOMP) > 0;
 }
 
 void INA228::setADCRange(bool flag)
 {
   uint16_t value = _readRegister(INA228_CONFIG, 2);
-  value &= ~0x0010;
-  if (flag) value |= 0x0010;
+  value &= ~INA228_CFG_ADCRANGE;
+  if (flag) value |= INA228_CFG_ADCRANGE;
   _writeRegister(INA228_CONFIG, value);
 }
 
 bool INA228::getADCRange()
 {
   uint16_t value = _readRegister(INA228_CONFIG, 2);
-  return value & 0x0010;
+  return (value & INA228_CFG_ADCRANGE) > 0;
 }
 
 
@@ -213,7 +245,7 @@ bool INA228::setMode(uint8_t mode)
 {
   if (mode > 0x0F) return false;
   uint16_t value = _readRegister(INA228_ADC_CONFIG, 2);
-  value &= 0x0FFF;  // ~(0x0F << 12);
+  value &= ~INA228_ADC_MODE;
   value |= (mode << 12);
   _writeRegister(INA228_ADC_CONFIG, value);
   return true;
@@ -222,14 +254,14 @@ bool INA228::setMode(uint8_t mode)
 uint8_t INA228::getMode()
 {
   uint16_t value = _readRegister(INA228_ADC_CONFIG, 2);
-  return (value >> 12) & 0x0F;
+  return (value & INA228_ADC_MODE) >> 12;
 }
 
 bool INA228::setBusVoltageConversionTime(uint8_t bvct)
 {
   if (bvct > 7) return false;
   uint16_t value = _readRegister(INA228_ADC_CONFIG, 2);
-  value &= 0xF1FF;  // ~(0x07 << 9);
+  value &= ~INA228_ADC_VBUSCT;
   value |= (bvct << 9);
   _writeRegister(INA228_ADC_CONFIG, value);
   return true;
@@ -238,14 +270,14 @@ bool INA228::setBusVoltageConversionTime(uint8_t bvct)
 uint8_t INA228::getBusVoltageConversionTime()
 {
   uint16_t value = _readRegister(INA228_ADC_CONFIG, 2);
-  return (value >> 9) & 0x07;
+  return (value & INA228_ADC_VBUSCT) >> 9;
 }
 
 bool INA228::setShuntVoltageConversionTime(uint8_t svct)
 {
   if (svct > 7) return false;
   uint16_t value = _readRegister(INA228_ADC_CONFIG, 2);
-  value &= 0xFE3F;  // ~(0x07 << 6);
+  value &= ~INA228_ADC_VSHCT;
   value |= (svct << 6);
   _writeRegister(INA228_ADC_CONFIG, value);
   return true;
@@ -254,14 +286,14 @@ bool INA228::setShuntVoltageConversionTime(uint8_t svct)
 uint8_t INA228::getShuntVoltageConversionTime()
 {
   uint16_t value = _readRegister(INA228_ADC_CONFIG, 2);
-  return (value >> 6) & 0x07;
+  return (value & INA228_ADC_VSHCT) >> 6;
 }
 
 bool INA228::setTemperatureConversionTime(uint8_t tct)
 {
   if (tct > 7) return false;
   uint16_t value = _readRegister(INA228_ADC_CONFIG, 2);
-  value &= 0xFFC7;  // ~(0x07 << 3);
+  value &= ~INA228_ADC_VTCT;
   value |= (tct << 3);
   _writeRegister(INA228_ADC_CONFIG, value);
   return true;
@@ -270,14 +302,14 @@ bool INA228::setTemperatureConversionTime(uint8_t tct)
 uint8_t INA228::getTemperatureConversionTime()
 {
   uint16_t value = _readRegister(INA228_ADC_CONFIG, 2);
-  return (value >> 3) & 0x07;
+  return (value & INA228_ADC_VTCT) >> 3;
 }
 
 bool INA228::setAverage(uint8_t avg)
 {
   if (avg > 7) return false;
   uint16_t value = _readRegister(INA228_ADC_CONFIG, 2);
-  value &= 0xF1FF;
+  value &= ~INA228_ADC_AVG;
   value |= avg;
   _writeRegister(INA228_ADC_CONFIG, value);
   return true;
@@ -286,7 +318,7 @@ bool INA228::setAverage(uint8_t avg)
 uint8_t INA228::getAverage()
 {
   uint16_t value = _readRegister(INA228_ADC_CONFIG, 2);
-  return value & 0x0010;
+  return (value & INA228_ADC_AVG);
 }
 
 
@@ -296,11 +328,23 @@ uint8_t INA228::getAverage()
 //
 int INA228::setMaxCurrentShunt(float maxCurrent, float shunt)
 {
-  if (maxCurrent > 10) return -1;
-  if (shunt < 0.005) return -2;
+  if (maxCurrent > 10) return -1;  //  TODO error code
+  if (shunt < 0.0001) return -2;   //  TODO error code
   _maxCurrent = maxCurrent;
   _shunt = shunt;
   _current_LSB = _maxCurrent * 1.9073486328125e-6;  //  pow(2, -19);
+
+  //  PAGE 31 (8.1.2)
+  float shunt_cal = 13107.2e6 * _current_LSB * _shunt;
+  //  depends on ADCRANGE in INA228_CONFIG register.
+  if (getADCRange() == 1)
+  {
+    shunt_cal *= 4;
+  }
+  //  shunt_cal must be written to REGISTER.
+  //  work in progress PR #7
+  _writeRegister(INA228_SHUNT_CAL, shunt_cal);
+
   return 0;
 }
 
@@ -314,6 +358,10 @@ float INA228::getShunt()
   return _shunt;
 }
 
+float INA228::getCurrentLSB()
+{
+  return _current_LSB;
+}
 
 ////////////////////////////////////////////////////////
 //
@@ -352,15 +400,25 @@ uint16_t INA228::getDiagnoseAlert()
 void INA228::setDiagnoseAlertBit(uint8_t bit)
 {
   uint16_t value = _readRegister(INA228_DIAG_ALERT, 2);
-  value |= (1 << bit);
-  _writeRegister(INA228_DIAG_ALERT, value);
+  uint16_t mask = (1 << bit);
+  //  only write new value if needed.
+  if ((value & mask) == 0)
+  {
+    value |= mask;
+    _writeRegister(INA228_DIAG_ALERT, value);
+  }
 }
 
-void INA228::clrDiagnoseAlertBit(uint8_t bit)
+void INA228::clearDiagnoseAlertBit(uint8_t bit)
 {
   uint16_t value = _readRegister(INA228_DIAG_ALERT, 2);
-  value &= ~(1 << bit);
-  _writeRegister(INA228_DIAG_ALERT, value);
+  uint16_t mask = (1 << bit);
+  //  only write new value if needed.
+  if ((value & mask ) != 0)
+  {
+    value &= ~mask;
+    _writeRegister(INA228_DIAG_ALERT, value);
+  }
 }
 
 uint16_t INA228::getDiagnoseAlertBit(uint8_t bit)
@@ -374,66 +432,81 @@ uint16_t INA228::getDiagnoseAlertBit(uint8_t bit)
 //
 //  THRESHOLD AND LIMIT REGISTERS 12-17
 //
+//  TODO - API ?
 
 void INA228::setShuntOvervoltageTH(uint16_t threshold)
 {
+  //  TODO ADCRANGE DEPENDENT
   _writeRegister(INA228_SOVL, threshold);
 }
 
 uint16_t INA228::getShuntOvervoltageTH()
 {
+  //  TODO ADCRANGE DEPENDENT
   return _readRegister(INA228_SOVL, 2);
 }
 
 void INA228::setShuntUndervoltageTH(uint16_t threshold)
 {
+  //  TODO ADCRANGE DEPENDENT
   _writeRegister(INA228_SUVL, threshold);
 }
 
 uint16_t INA228::getShuntUndervoltageTH()
 {
+  //  TODO ADCRANGE DEPENDENT
   return _readRegister(INA228_SUVL, 2);
 }
 
 void INA228::setBusOvervoltageTH(uint16_t threshold)
 {
   if (threshold > 0x7FFF) return;
+  //float LSB = 3.125e-3;
   _writeRegister(INA228_BOVL, threshold);
 }
 
 uint16_t INA228::getBusOvervoltageTH()
 {
+  //float LSB = 3.125e-3;
   return _readRegister(INA228_BOVL, 2);
 }
 
 void INA228::setBusUndervoltageTH(uint16_t threshold)
 {
   if (threshold > 0x7FFF) return;
+  //float LSB = 3.125e-3;
   _writeRegister(INA228_BUVL, threshold);
 }
 
 uint16_t INA228::getBusUndervoltageTH()
 {
+  //float LSB = 3.125e-3;
   return _readRegister(INA228_BUVL, 2);
 }
 
 void INA228::setTemperatureOverLimitTH(uint16_t threshold)
 {
+  //float LSB = 7.8125e-3;  //  milliCelsius
   _writeRegister(INA228_TEMP_LIMIT, threshold);
 }
 
 uint16_t INA228::getTemperatureOverLimitTH()
 {
+  //float LSB = 7.8125e-3;  //  milliCelsius
   return _readRegister(INA228_TEMP_LIMIT, 2);
 }
 
 void INA228::setPowerOverLimitTH(uint16_t threshold)
 {
+  //  P29
+  //  Conversion factor: 256 × Power LSB.
   _writeRegister(INA228_POWER_LIMIT, threshold);
 }
 
 uint16_t INA228::getPowerOverLimitTH()
 {
+  //  P29
+  //  Conversion factor: 256 × Power LSB.
   return _readRegister(INA228_POWER_LIMIT, 2);
 }
 
@@ -442,7 +515,7 @@ uint16_t INA228::getPowerOverLimitTH()
 //
 //  MANUFACTURER and ID REGISTER 3E/3F
 //
-bool INA228::getManufacturer()
+uint16_t INA228::getManufacturer()
 {
   uint16_t value = _readRegister(INA228_MANUFACTURER, 2);
   return value;
@@ -451,13 +524,13 @@ bool INA228::getManufacturer()
 uint16_t INA228::getDieID()
 {
   uint16_t value = _readRegister(INA228_DEVICE_ID, 2);
-  return (value >> 4) & 0xFFF0;
+  return (value >> 4) & 0x0FFF;
 }
 
 uint16_t INA228::getRevision()
 {
   uint16_t value = _readRegister(INA228_DEVICE_ID, 2);
-  return value & 0x0F;
+  return value & 0x000F;
 }
 
 
@@ -473,7 +546,7 @@ uint32_t INA228::_readRegister(uint8_t reg, uint8_t bytes)
 
   _wire->requestFrom(_address, (uint8_t)bytes);
   uint32_t value = 0;
-  for ( int i = 0; i < bytes; i++)
+  for (int i = 0; i < bytes; i++)
   {
     value <<= 8;
     value |= _wire->read();
@@ -482,15 +555,15 @@ uint32_t INA228::_readRegister(uint8_t reg, uint8_t bytes)
 }
 
 
-float INA228::_readRegisterF(uint8_t reg, uint8_t bytes)
+double INA228::_readRegisterF(uint8_t reg, uint8_t bytes)
 {
   _wire->beginTransmission(_address);
   _wire->write(reg);
   _wire->endTransmission();
 
   _wire->requestFrom(_address, (uint8_t)bytes);
-  float value = 0;
-  for ( int i = 0; i < bytes; i++)
+  double value = 0;
+  for (int i = 0; i < bytes; i++)
   {
     value *= 256.0;
     value += _wire->read();